Move cells into a single ArrayHashMap.
Previously, cells were stored in an ArrayHashMap in a Column struct, which were in turn stored in
an ArrayHashMap in a Sheet struct. Storing all the cells in the same map allows for better cache
locality across rows, and the ability to influence cache locality by simply sorting the map. It
also removes the extra level of indirection for each cell access, and maps Positions directly to
Cells rather than mapping x and y separately. Columns still exist in an ArrayHashMap in Sheet to
store per-column precision and width values.