Make rendering more granular
Previously, the entire screen was cleared and redrawn after every input. This was inefficient and
led to flickering on some terminals such as the Linux tty. Rendering is now done without clearing
the screen, updating individual elements as needed. In particular, redraws of the cell contents
are only done when the cell contents change or the view scrolls. This fixes the issue of flicker
on certain terminals. Also updates spoon to the newest version.