Fix memory leak, add cell delete, re-use ASTs
Fixes a memory leak caused by not freeing old cell ASTs when loading a new file. ASTs are now
stored in a list in ZC when they are deleted, and are re-used from this list when a new AST is
required. Cells can now be deleted, and their ASTs go into the afformentioned list to be re-used.