Bounds check cell addresses, many UI fixes
Cell addresses are now bounds checked. Attempting to use a column past CRXP or a row above 65535
will now give an error. Increased the number of columns by 1, as we used to erroneously have only
65534 columns. This was because the code that translated column address strings into column indices
started from 1, when the column indices started at 0. Column headings will now be redrawn when
the size of the row number column changes. Fixes many integer overflow issues related to handling
UI drawing when at the last possible column/row. Fixes drawing artifacts in the remaining space
when screen is at the last column. Fixes overflow issuse related to drawing over the old cursor
position when the cursor is moved. Fixes the last column not being drawn.

Also adds 2 new key bindings in normal mode: g and G to go to the first and last cells in a column
respectively.