Overhaul input handling
Input handling now uses a critbit tree for mapping keys to actions. It maps key-sequence strings
to enums that are switched on. Critbit trees allow us to easily test if a string is a prefix of
any string in the tree, and forbids storing prefixes of keys. This allows us to easily tell if an
input sequence can become a valid mapping. The mapped key-sequences strings are human readable,
e.g. control-k is '<C-k>', escape is '<Escape>', etc. This will allow for easier runtime key
mapping in the future. This approach to handling keys is inherently slower than the previous
method, which was just a series of switch statements. This method requires heap allocating tree
nodes and traversing the tree to resolve key inputs. This is worth it for the flexibility it
offers.

This commit adds new inside/around motions for operator pending modes. (, [, {, ", ', `, word, and
WORD are all supported.