Implement command to/until motions, update spoon
Implements to/until motions in forward and backward directions. These motions move the cursor to or
just before/after the next/previous instance of the next inputted character. Works with counts and
operator pending modes.

Update spoon to lastes, which fixes a bug related to cutting off codepoints that appear at the end
of the input buffer.