Heavily optimize keymap memory usage
Not particularly needed but I wanted to do it. ~21KB to ~4KB. Does some absolutely cursed
optimizations, such as storing 2 unicode chars in a [7]u8 and duplicating the fields of the Motion
tagged union into the CommandAction one. Fixes a bug in the critbit tree implementation that caused
a compile error for non-slice keys, and a runtime error for slice keys that weren't u8. Reduces the
memory overhead of critbit nodes to 8 bytes per node.