Initial support for text in cells
Adds support for text in cells. Changes the syntax for assigning cell
numbers to 'let A0 = 0', and the syntax for assigning cell text is
'label a0 = "text"'.

String literals are supported in label assignment as text surrounded in
single or double quotes. The text value of a cell can be referenced via
its cell address. Strings can be concatenated via the new '#' operator.

Attempting to use numerical literals/operators in a label assignment
and vice-versa will produce a parse error.

Cells can contain both numerical and string values at the same time.
Only the text will be shown. In the future it will be possible to
reference a cells numerical value in the text.