Many fixes
Fix crash when executing a set_text_cell undo.
Fix crash when changing width of a column with a text cell.
Fix Sheet.set_width not appending an undo group_end marker.
Fix colours not applying to the last character in a cell.
Fix centering and truncation of non-ASCII text.
Motions that move to the first/last cell in a column/row now account
for text cells as well as numerical cells.
Add 'aa' keybind in normal mode to expand the width of a column to fit
its cells.