Add @upper and @lower builtins for strings
Adds 2 new builtin functions, @upper and @lower, for use with
strings. @upper converts a single argument to an upper-case
representation, and @lower converts a single argument to a lower-case
representation. Currently only works for ASCII, though libgrapheme will
be used in the future for unicode. These functions can only be used
in a string context, so currently only when assigning a label.

This commit also adds 2 new error messages for using numeric builtins
in a string context, and for using string builtins in a numeric
context.