Many eval improvements & fixes
Due to the fact that we cache every eval result, sorting the cells by
topological order before evaling wasn't needed. This has been removed,
reducing debug mode eval time by 60% on example.zc.

NaN values are now handled better. Previously, NaN meant an evaluation
error. It was also used as a sort of `null` value for cells that hadn't
been evaluated yet during evaluation. This meant that if you set the
value of a cell to 0 / 0 (NaN), the cell's value would never be cached.
If you had lots of NaNs this could cause a big slowdown when evaling.
This is now fixed. NaNs no longer print an "ERROR" to a cell, instead
printing "nan".

Fixes cyclical reference detection for numeric values, which wasn't
working as intended. It was only detecting cycles with a depth of 0,
but now works for all cycles.

Adds cyclical reference detection for string values.

More granular evaluation is planned soon, and (hopefully) should not
be too difficult to implement.