Optimize R-Tree performance
Still need to merge getSingle and getOrPut

Optimizes the getSingle function so we check if a node contains the key
before calling the function.
Found optimal values for min_children in the R-Trees.
Reduced memory usage and allocation time for R-Tree branch nodes by
using a *BoundedArray and memory pool.