Add ref nodes to ASTs, rename Range to Rect
References are positions that hold pointers to specific rows and columns
in a sheet. This allows references to be automatically adjusted when
rows and columns are deleted in a sheet.

Ranges are now over references instead of positions, and the old Range
struct has been renamed to Rect.

Also adds a function to convert positions in an AST to references. This
function is currently not used.