Refactor AST functions
Refactors AST functions to take a list of nodes along with a root node
index. This is to allow for a future refactor of AST handling where all
AST nodes will be stored in a single MultiArrayList in a Sheet, and
cells will hold only a u32 index into the node list.