Refactor AST memory handling
All AST nodes are now stored in a list in Sheet. Cells store the index
of the root node of their expression.

This way of handling the memory allows for a number of cool things.
First is that you can reserve memory for a ton of ASTs with just a
single allocation, which would vastly improve bulk creation of cells
when implemented.

Second, the memory overhead for storing ASTs is reduced by 75%.
Previously each cell stored a MultiArrayList containing its AST nodes,
which consisted of a pointer and a u32 len + capacity. Cells now store
only a u32 index into the ast_nodes array.

Third is that it makes serialisation significantly easier and faster. We
will be able to serialise all ASTs by just writing the contents of the
ast_nodes array to disk. The previous implementation would have required
traversing every cell, and somehow mapping each cell to its ast nodes.
Cells now store a u32 index which significantly simplifies serialisation
also.

One drawback of this approach is that it makes it impossible to change
ASTs after they are entered into the sheet. It also means ASTs cannot be
freed. This is fine however, as we keep all the ASTs around anyway for
undo support.