Implement more commands, fix bugs, add more tests
Fixes bug where cells would be added to the cell tree twice. This was
because the RTree put function only checked if the key existed in the
most optimal leaf node, while there can be multiple candidates for where
the key exists. This bug only manifested visibly when using functions
on ranges as they use the cell tree to look up cells in the ranges
passed to them. This wasn't an issue for the DependentTree as all put
operations go through putSlice which did the proper check.

Fixes a bug where the top left and bottom right references in a range
referenced by a cell expression would be redundantly added as
dependencies of the cell. This bug shouldn't have caused any unexpected
behaviour, but would add redundant nodes to the dependent tree.

Fixes a bug where the fill command did not pre-allocate enough memory.
This was causing an assertion trip since the memory for AST nodes was
refactored.

Introduces more commands for sheet operations. This includes :undo,
:redo, :delete, for normal usage. Also added are commands for asserting
the properties of the sheet for unit tests, which are only available in
debug builds.

Implements more test cases as .zc files using the new test commands.
These test cases caught the aformentioned bugs.