Parser: Improve tokenizing and parsing performance
Measures a ~24% decrease in the time it takes to parse 1 million
statements. The input to the tokenizer is now null terminated and it
uses labeled continues instead of a while loop. This results in a ~60%
decrease in branch misses. Tokenization is now done seperately from
parsing, which presumably decreases pressure on the instruction cache.
Also reduced the number of branches in parseAddExpr and parseMulExpr.

About half of the performance win comes from making the *Parser
argument noalias in Parser.addNode.