Ditch row/column references, improve flatlist perf
FlatListPool performance is now massively improved. Appending to a list
previously caused a loop through every other list. The new version now
has an individual capacity for every list, and if a list exceeds it's
capacity it's contents are moved to the end of the buffer, which means
other lists don't have to have their offsets updated. This substantially
improves file loading performance.

Row and column references have been ditched in favour of the previous
implementation of using raw u32 positions. This increases performance
and simplifies the implementation, at the cost of slower row/column
insertion/deletions (which are not implemented yet.)