Sheet: 35x speed up in sheet loading
Eliminates two comically bad bottlenecks. A previous commit changed
parsing and tokenizing from being interleaved to being pipelined, which
mean all tokens were collected in an array and given to the parser. This
was a good idea in theory, but the parser worked only one single lines,
which meant that this token array was being alloced/freed for every
single line in a file. Changing this to tokenize the whole file results
in a large speed up.

Another bottleneck was in the growth rate of `FlatListPool`. Basically
it didn't over-allocate nearly enough and resulted in many hits to the
allocator, which is now fixed.

Also includes a few other performance improvements for file loading,
like pre-allocating more memory and batching the steps required for
inserting cells instead of interleaving them to improve cache
performance.

Replaces the R-Tree implementations with PH Trees. This results in
significantly improved insert/delete performance as compared to R-Trees.
The impact on search performance is negligible.