PH tree: don't create redundant branch nodes
Updates the PH tree implementation to allow any branch node to store
value/branch nodes as children. In the previous implementation value
nodes were only stored as children by branch nodes with a postfix
length of one. This change reduces the amount of nodes in the tree in
most cases including reducing the depth of the tree by 1. This also
greatly simplifies the logic for the remove function, which now only
affects two branch nodes at most. This is also a prerequisite to writing
a multithreaded implementation.