ZC: Implement text alignment
Implements basic text alignment, allowing for cells with string values
to have their text displayed as left, right, or center aligned. The
default remains as center aligned. Text alignment can be set with the
text-align command, or by the `<`, `>`, and `|` keybinds in normal or
visual mode.

Text alignment is not currently undoable and text alignment state is not
serialized. This will be added later.