Tui: Refactor Tui code, fix some bugs
Overhauls a lot of the TUI code to be cleaner and deduplicates some
logic.

Fix bug where moving the cursor from Lua would not correctly update the
selected column/row heading.

Fix assertion trips when the screen is at the right edge of the sheet.

Fix integer overlow when two columns are >= 2^16 units apart