UI: Implement themes, fix some bugs
The terminal UI now supports loading user-defined themes at runtime.
The system for loading themes is UI-independent, and this commit has
made steps to begin decoupling the UI frontend from the rest of the
application. The goal of this is to make it easy to use any UI frontend
with cellulator in the future.

To support themes, the UI frontend exposes a few things. A UI
name, a file extension string for theme files, and a function
to set a theme given a path to a theme file. The definition of a
theme file is dependent on the UI frontend. For example, a web UI
might store themes as CSS files. The current terminal UI defines
themes with lua files. Theme files are stored in the format
`$XDG_CONFIG_HOME/cellulator/themes/$UI_NAME/$THEME_NAME.$UI_FILE_EXT`.
For example, a TUI theme called `mytheme` would be stored in
`$XDG_CONFIG_HOME/cellulator/themes/terminal/mytheme.lua`.

The current theme can be set with the new `set` command. It takes two
arguments, the property to set and the value to set it to. Two
properties are currently supported, being `theme` and `truecolor`. The
`truecolor` property is a boolean property, so calling `set truecolor`
without specifying a value will set it to true.

The `unset` command has also been implemented. In the case of themes it
will restore the default theme, and calling `unset` with a boolean
property will set it to false. This may be subject to change in the
future.

The rationale for the truecolor property is that many terminals that
support truecolor (24 bit color) don't properly advertise this
capability, so in those cases the user can override the value and force
enable truecolor. Users can also disable truecolor if they wish.