Fix bug in parser and sheet load
Fixes hitting an unreachable statement in the parser where the string
data for a number token would contain invalid text, which would get
passed to parseFloat.

Loading a file that set the same cell twice would cause an assertion
trip due to the interpretSource function assuming that the same cell
wouldn't be written to more than once.