Add fuzzing step, fix tokenizer bug
Fixes a bug in the tokenizer where a single . character would result in
a number token, which would cause UB when parsed.

Adds a fuzzing step to build.zig. It uses the system installed afl++
instead of zig-afl-kit because that package doesn't work on my system
and I don't care to debug it.