Fix unicode handling in command mode
Text cursor in command line now does proper grapheme cluster
segmentation along with setting the cursor based on the correct display
width of the text. The display width of the text is dependent on the
terminal emulator. The cursor is set using the terminals save/restore
cursor functionality and not with zg's text width. The point on the
screen at which the command buffer scrolls is still determined using
zg's text width mechanisms so for a small amount of grapheme clusters
this may not align with the text width reported by the terminal
emulator. However, this is fine, with the benefit being that the cursor
will always visually line up with the text regardless of rendering
differences between terminal emulators.