Improve render perf of string cells, fix TUI bug
Fixes a bug in the TUI where column headings weren't being truncated
properly.

Improves performance of rendering string cells. Calculating the display
width of the strings was required to align them correctly, but the old
implementation would iterate the entire string to calculate the width,
which for long strings could be costly. The new implemenation takes a
max width parameter and will exit early if the string exceeds that max
width.