Implement multiple sheets
Multiple sheets can now exist at once. Adds a line at the bottom of the
screen detailing the currently opened sheets. Opening a new file opens
the file in a new sheet.

An empty sheet can be created by omitting the filepath argument to the
`e` command. Opening files with the `e` command always opens the file
in a new sheet. The `sheet-close` command and the <C-w>q keybind closes
the currently selected sheet. The `gn` and `gp` keybinds go the next
and previous sheets respectively. The `sheet-rename` command renames the
current sheet to the given argument if only one argument is specified.
If two arguments are specified the first argument is the name of the
sheet to rename and the second argument is the new name for that sheet.
Sheet names cannot be an empty string, or conflict with the names of
other existing sheets. Sheet names are separate from the underlying
filepath of the sheet, which is still display in the top status line.

Also fixes a bug where the left bracket of the cell expression indicator
in the status line would be incorrectly styled with the count style.