Fix TUI bug, reduce number of bytes sent to term
Fixes a bug where the the column/row headings for the previously
selected cell weren't being unhighlighted if the previous cursor
location is off screen.

Removes some cases where the selected column/row headings were being
written twice in a single render pass. This reduces the number of
bytes sent to the underlying terminal emulator and improves rendering
performance.

Eventually it is planned for Shovel (the TUI library in use) to support
screen buffering and diffing, which would change the terminal UI to an
immediate mode UI, rather than being retained mode due to the inherent
nature of terminal emulators. This would almost certainly reduce the
number of small UI bugs that keep popping up.