Add syntax highlighting, fix input bug
Adds syntax highlighting for statements. This highlighting shows up in
the command line and for the expression of the selected cell in the
status line. Several new theme keys have been added to adjust the
highlighting styles.

Fixes a long standing input bug where typing a < character on the
command line would insert <<. This was because < is used to delimit a
modifier sequence for keybinds. The function for converting input into
the readable format to look up keybinds was also being used to write
readable inputs to the command line.