Fix redrawing cells when not necessary
Cells were incorrectly being redrawn after every input. This was because
the function to clear the input buffer after an unambiguous input would
need to redraw the cells if there was the input hint dialogue on screen.
The check for whether the last input was ambiguous was erroneous and
was always hit.