Add completions for commands, small fixes
The completion system has been reworked to be generic instead of only
working for filepaths. Completions for command names has now been added.

Fixes leading whitespace not rendering in the command line.

Fixes a divide by zero when rendering completions in a tiny terminal.

Selected completion is now styled to the end of the line, rather than
just the end of the text.

Renames some TUI elements.

Reorganized command handling into a namespace.