Implement cell references and ranges as values
Cell references and ranges are now values in ZC. They can be stored in
cells like any other value. Cell references can be created with the &
operator, and can be dereferenced with the * operator. A cell literal
can now either evaluate to the value of the cell or a cell reference,
depending on the context.

The `:` syntax is no longer special cased and is now an operator. The
left and right operands must evaluate to cell references. Anything other
than cell literals or references to cell literals as the operands will
create a dynamic range, accesses through which are volatile.

Builtins which access cell values evaluate their arguments in a
reference context, meaning cell literals will be interpreted as
references. These functions will also automatically dereference their
arguments once and once only.

Accesses through a dynamic reference make the expression volatile. Cells
with volatile expressions are updated on every change to the sheet.

Adds type tracking to the parser, which will be used in the future for
better error messages on type mismatches.

Cell references and ranges have a truthiness value of true.

Styles for different cell types are now properly used. Adds styles for
cell reference and range types.

Fixes a bug where old dependencies were not removed when a cell's
expression was changed.