Make evaluation iterative, remove ast node fields
Makes expression evaluation iterative using a heap allocated stack,
instead of being recursive. This removes the possibility of a stack
overflow during evaluation.

ASTs now have a special node that marks the end of an AST. It has the
index of the root node + 1, and contains the number of nodes in the
AST. This lets us find the left-most/first node, and the number of nodes
in the AST without iterating through it. This is useful for reserving
memory and evaluating left to right from a root node index.

Binary operator AST nodes no longer store references to their data, as
the indices of the child nodes can be calculated from the array order.