Implement AST abstraction
The list of AST nodes is now stored in a struct along with the strings
required by those AST nodes. Various AST functions are now methods on
the AST structure.

Adds a `MultiList` implementation, which is essentially the same as
`std.MultiArrayList(T).Slice` but with more methods. It's no longer
required to manually convert to/from MultiArrayList in order to resize
or modify the list. Additionally, it integrates a distinct Index type
and has methods for working with these type safe indexes. This vastly
reduces the number of conversions between index types and integers,
improving type safety.

Overall, the API and its uses are much cleaner and more maintainable.