Refactor parser API, append strings at parse time
Strings are now appended to the string buffer in the AST during parse
time, instead of as a separate pass after parsing. This simplifies call
sites and should perform better.

All single-shot parsing is now done through a single function, instead
of by initializing a parser and calling a method. Initializing a parser
is now only necessary when you want to parse multiple statements.