Remove isSingle from AST, clean up some printing
isSingle made no sense and has been properly replaced with a check on
the precedence of the involved operators.

The range operator now has a correct precedence and will surround its
arguments in parentheses if necessary. Currently there is no way to
yield a cell reference from an operator with a lower precedence, so this
doesn't actually affect anything right now.

The power operator now prints without spaces surrounding it.