Fix panic when taking an invalid reference
The only valid argument to the reference-of operator (&) is a cell
literal. The interpreter incorrectly assumed that taking the reference
of anything other than the cell literal would not pass the parser, which
wasn't the case. The parser now correctly rejects taking a reference to
anything other than a cell literal.